source-code/dvpa-api/requirements.txt
aniso8601==8.1.0
appnope==0.1.2
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==6.6
decorator==4.4.2
Flask==1.1.2
Flask-JWT-Extended==3.25.0
flask-mongoengine==0.8
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
flask-peewee==0.6.7
flask-restplus==0.13.0
Flask-Script==2.0.5
Flask-WTF==0.12
gunicorn==20.0.4
idna==2.10
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==0.23
mongoengine==0.10.6
mysqlclient==2.0.1
parso==0.7.1
peewee==2.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
PyJWT==1.7.1
pymongo==3.3.0
PyMySQL==0.10.1
pyrsistent==0.17.3
pytz==2020.5
PyYAML==5.1
requests==2.25.1
six==1.15.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==0.16.1
wtf-peewee==0.2.6
WTForms==2.1